Animation order: only show valid options
