+++ assets/src/stores/amp-story/index.js
@@ -22,6 +22,7 @@ export default registerStore(
 		selectors,
 		actions,
 		initialState: {
+			animationOrder: {},
 			reordering: {
 				blockOrder: [],
 				isReordering: false,
