+++ assets/src/stores/amp-story/selectors.js
@@ -26,6 +26,44 @@ export function getAnimationPredecessor( state, page, item ) {
 }
 
 /**
+ * Determines whether a given predecessor -> item combination is valid.
+ *
+ * Returns false if the predecessor itself is not animated or if it would result in a cycle.
+ *
+ * @param {Object} state       Editor state.
+ * @param {string} page        ID of the page the item is in.
+ * @param {string} item        ID of the animated item.
+ * @param {string} predecessor ID of the animated item's predecessor.
+ *
+ * @return {boolean} True if the animation predecessor is valid, false otherwise.
+ */
+export function isValidAnimationPredecessor( state, page, item, predecessor ) {
+	if ( undefined === predecessor ) {
+		return true;
+	}
+
+	const pageAnimationOrder = state.animationOrder[ page ] || [];
+	const predecessorEntry = pageAnimationOrder.find( ( { id } ) => id === predecessor );
+
+	const hasCycle = ( a, b ) => {
+		let parent = b;
+
+		while ( parent !== undefined ) {
+			if ( parent === a ) {
+				return true;
+			}
+
+			const parentItem = pageAnimationOrder.find( ( { id } ) => id === parent );
+			parent = parentItem ? parentItem.parent : undefined;
+		}
+
+		return false;
+	};
+
+	return predecessorEntry && ! hasCycle( item, predecessor );
+}
+
+/**
  * Returns the currently selected page.
  *
  * @param {Object} state Editor state.
